Churn: Rewrote churn mechanism, now using home-built
Instead of the Lifecycle feature in INET, use own solution. INET's code will throw an error if an UDP message is received by an offline host, which will eventually happen to us… So this is a first try of an homemade online/offline-mechanism that simply discards UDP packets when offline.