Introduce PING messages for active offline detection
Sends out PINGs at given time intervals, if no other communication to the peer in question happens in-between; in that case, the timer for the next PING is reset, i.e. PINGs are sent out after <x> seconds of silence. PINGs are only acknowledged (RCKACK), no specific answer is sent.